Switch to using the aiohttp library rather than requests (#2)
Problem: We need to run send price levels and RFQs in parallel for each
individual request. This is done so that we get the exact levels from
prod at RFQ time which are then used to compute price deviation. However
because we're using the synchronous `requests` module, this is not
possible.

Solution: Switch to `aiohttp` which is async.